+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Next
 * Add pycryptodomex version requiremenet range (`>=3.9,<4`)
 * Fix using bytes file paths for `CDNReader` and `SDFilesystem`
+* Support auto-closing underlying file for `CTRFileIO` and `CBCFileIO`
 
 ## v0.4.5 - October 24, 2020
 * Fix loading RomFS from a filename in `RomFSReader`